backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
langchain
langchain-google-genai
supabase
numpy
sentence-transformers
pdfplumber
groq
